# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Crypto Backtester API
  version: 0.8.0
  description: |
    Chart client API — unified chart-data, historical candles, indicator
    compute, structured symbols, users, watchlists, replay (WebSocket v2),
    backtest HTTP, screener scan, AI NL→DSL, and live candle WS.

    **Base path (REST):** `/api/v1`

    **Authentication:** none. Optional `user_id` on create bodies is metadata only.
    Resources are open by UUID. `/auth/*` and `/ws/tickets` have been removed.

    **Timestamps:** Candle `time`, replay `start`/`end`/`cursor`, backtest
    windows, scan windows, and indicator `from`/`to` use **Unix seconds UTC**.

    **Errors (REST):** JSON envelope `{"error": {"code": "...", "message": "..."}}`.

    **WebSocket:** Documented under `x-websocket` (not part of OpenAPI path model).
  contact:
    name: crypto-backtester
  license:
    name: Unspecified
servers:
  - url: http://localhost:8000
    description: Local uvicorn (`python -m api` or `uvicorn api.main:app`)
...
paths:
...
  /api/v1/users/{user_id}/watchlists:
    get:
      tags: [watchlists]
      summary: List watchlists
      operationId: listWatchlists
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Watchlists with ordered symbols
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistResponse"
        "401":
          description: Missing or invalid JWT
        "403":
          description: JWT subject does not match user_id
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [watchlists]
      summary: Create watchlist
      operationId: createWatchlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistCreate"
      responses:
        "201":
          description: Created watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/users/{user_id}/watchlists/{watchlist_id}:
    get:
      tags: [watchlists]
      summary: Get watchlist (owner-only)
      operationId: getWatchlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/WatchlistId"
      responses:
        "200":
          description: Watchlist detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [watchlists]
      summary: Update watchlist metadata (owner-only)
      operationId: updateWatchlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/WatchlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistUpdate"
      responses:
        "200":
          description: Updated watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [watchlists]
      summary: Delete watchlist (owner-only)
      operationId: deleteWatchlist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/WatchlistId"
      responses:
        "204":
          description: Watchlist deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user_id}/watchlists/{watchlist_id}/symbols:
    put:
      tags: [watchlists]
      summary: Replace watchlist symbols (owner-only)
      operationId: replaceWatchlistSymbols
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/WatchlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistSymbolsUpdate"
      responses:
        "200":
          description: Watchlist with new symbol order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/auth/register` or `/auth/login`.
  parameters:
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    WatchlistId:
      name: watchlist_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing:
              value:
                error:
                  code: UNAUTHORIZED
                  message: "Unauthorized"
            invalid:
              value:
                error:
                  code: INVALID_TOKEN
                  message: "Invalid or expired token"
    Forbidden:
      description: Authenticated user is not the owner of the target resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            not_owner:
              value:
                error:
                  code: FORBIDDEN
                  message: "Cannot access another user's resources"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            symbol:
              value:
                error:
                  code: SYMBOL_NOT_FOUND
                  message: "Unknown symbol: FOO/USDT"
            user:
              value:
                error:
                  code: USER_NOT_FOUND
                  message: "Unknown user: ..."
            replay:
              value:
                error:
                  code: REPLAY_NOT_FOUND
                  message: "Unknown replay session: ..."
            scan:
              value:
                error:
                  code: SCAN_NOT_FOUND
                  message: "Scan not found"
            backtest:
              value:
                error:
                  code: BACKTEST_NOT_FOUND
                  message: "Backtest run not found"
            ai_session:
              value:
                error:
                  code: SESSION_NOT_FOUND
                  message: "Unknown or expired session"
    ValidationError:
      description: Validation or business rule failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            timeframe:
              value:
                error:
                  code: INVALID_TIMEFRAME
                  message: "Unsupported timeframe: 2m"
            range:
              value:
                error:
                  code: INVALID_RANGE
                  message: from must be <= to
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
...
    WatchlistCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        symbols:
          type: array
          items:
            type: string
          default: []
    WatchlistUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
        is_default:
          type: boolean
          nullable: true
        sort_order:
          type: integer
          nullable: true
    WatchlistSymbolsUpdate:
      type: object
      required: [symbols]
      properties:
        symbols:
          type: array
          items:
            type: string
          description: Ordered symbol list (active symbols only)
    WatchlistResponse:
      type: object
      required: [id, user_id, name, is_default, sort_order, symbols, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        is_default:
          type: boolean
        sort_order:
          type: integer
        symbols:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...